refactor llmapi e2e tests
