+++ tests/integration/defs/test_e2e.py
@@ -1056,216 +1056,6 @@ def test_mistral_large_hidden_vocab_size(llama_example_root, llm_venv,
 run_llm_path = os.path.join(os.path.dirname(__file__), "_run_llmapi_llm.py")
 
 
-@pytest.mark.parametrize("model_name,model_path", [
-    ("llama", "llama-models/llama-7b-hf"),
-    ("llama", "codellama/CodeLlama-7b-Instruct-hf"),
-])
-def test_llmapi_load_engine_from_build_command(llm_root, llm_venv, engine_dir,
-                                               model_name, model_path):
-    llama_example_root = os.path.join(llm_root, "examples", "models", "core",
-                                      model_name)
-    dtype = 'float16'
-    cmodel_dir = os.path.join(engine_dir, f"{model_name}-engine")
-
-    ckpt_dir = convert_weights(llm_venv=llm_venv,
-                               example_root=llama_example_root,
-                               cmodel_dir=cmodel_dir,
-                               model=model_name,
-                               model_path=f'{llm_models_root()}/{model_path}',
-                               data_type=dtype)
-
-    engine_dir = os.path.join(engine_dir, f"{model_name}-engine")
-
-    build_cmd = [
-        "trtllm-build",
-        f"--checkpoint_dir={ckpt_dir}",
-        f"--output_dir={engine_dir}",
-        f"--max_batch_size={8}",
-        f"--max_input_len={924}",
-        f"--max_seq_len={1024}",
-        f"--max_beam_width={1}",
-        f"--gemm_plugin={dtype}",
-        f"--gpt_attention_plugin={dtype}",
-    ]
-    check_call(" ".join(build_cmd), shell=True, env=llm_venv._new_env)
-
-    venv_check_call(llm_venv, [
-        run_llm_path,
-        "--model_dir",
-        engine_dir,
-    ])
-
-
-@pytest.mark.parametrize("model_name,model_path", [
-    ("llama", "llama-models-v2/llama-v2-7b-hf"),
-])
-def test_llmapi_load_engine_from_build_command_with_lora(
-        llm_root, llm_venv, engine_dir, model_name, model_path):
-    llama_example_root = os.path.join(llm_root, "examples", "models", "core",
-                                      model_name)
-    dtype = 'bfloat16'
-    cmodel_dir = os.path.join(engine_dir, f"{model_name}-engine")
-
-    ckpt_dir = convert_weights(llm_venv=llm_venv,
-                               example_root=llama_example_root,
-                               cmodel_dir=cmodel_dir,
-                               model=model_name,
-                               model_path=f'{llm_models_root()}/{model_path}',
-                               data_type=dtype)
-
-    engine_dir = os.path.join(engine_dir, f"{model_name}-engine")
-
-    build_cmd = [
-        "trtllm-build",
-        f"--checkpoint_dir={ckpt_dir}",
-        f"--output_dir={engine_dir}",
-        f"--max_batch_size={1}",
-        f"--max_input_len={924}",
-        f"--max_seq_len={1024}",
-        f"--max_beam_width={1}",
-        f"--gemm_plugin={dtype}",
-        f"--gpt_attention_plugin={dtype}",
-        f"--lora_plugin={dtype}",
-        f"--lora_target_modules=attn_q",
-    ]
-    check_call(" ".join(build_cmd), shell=True, env=llm_venv._new_env)
-
-    venv_check_call(llm_venv, [
-        run_llm_path,
-        "--model_dir",
-        engine_dir,
-    ])
-
-
-@pytest.mark.parametrize("model_name,model_path", [
-    ("llama", "llama-models-v2/TinyLlama-1.1B-Chat-v1.0"),
-])
-def test_llmapi_build_command_parameters_align(llm_root, llm_venv, engine_dir,
-                                               model_name, model_path):
-    from tensorrt_llm.llmapi import LLM
-    from tensorrt_llm.llmapi.llm_utils import BuildConfig
-    llama_example_root = os.path.join(llm_root, "examples", model_name)
-    dtype = 'float16'
-    cmodel_dir = os.path.join(engine_dir, f"{model_name}-engine")
-
-    ckpt_dir = convert_weights(llm_venv=llm_venv,
-                               example_root=llama_example_root,
-                               cmodel_dir=cmodel_dir,
-                               model=model_name,
-                               model_path=f'{llm_models_root()}/{model_path}',
-                               data_type=dtype)
-
-    engine_dir = os.path.join(engine_dir, f"{model_name}-engine")
-
-    build_cmd = [
-        "trtllm-build",
-        f"--checkpoint_dir={ckpt_dir}",
-        f"--output_dir={engine_dir}",
-        f"--max_batch_size={4}",
-        f"--max_input_len={111}",
-        f"--max_seq_len={312}",
-        f"--max_beam_width={4}",
-        f"--gemm_plugin={dtype}",
-        f"--gpt_attention_plugin={dtype}",
-    ]
-    check_call(" ".join(build_cmd), shell=True, env=llm_venv._new_env)
-
-    build_config = BuildConfig()
-    # change some building parameters
-    build_config.max_batch_size = 4
-    build_config.max_beam_width = 4
-    build_config.max_input_len = 111
-    build_config.strongly_typed = True
-    build_config.max_seq_len = 312
-    build_config.plugin_config._gemm_plugin = dtype
-    build_config.plugin_config._gpt_attention_plugin = dtype
-
-    llm = LLM(model=f'{llm_models_root()}/{model_path}',
-              build_config=build_config)
-    tmpdir = tempfile.TemporaryDirectory()
-    llm.save(tmpdir.name)
-    build_cmd_cfg = None
-    build_llmapi_cfg = None
-
-    with open(os.path.join(engine_dir, "config.json"), "r") as f:
-        engine_config = json.load(f)
-
-        build_cmd_cfg = BuildConfig.from_dict(
-            engine_config["build_config"]).to_dict()
-
-    with open(os.path.join(tmpdir.name, "config.json"), "r") as f:
-        llm_api_engine_cfg = json.load(f)
-
-        build_llmapi_cfg = BuildConfig.from_dict(
-            llm_api_engine_cfg["build_config"]).to_dict()
-
-    assert build_cmd_cfg == build_llmapi_cfg
-
-
-def test_llmapi_load_ckpt_from_convert_command(llm_root, llm_venv, engine_dir):
-    llama_example_root = os.path.join(llm_root, "examples", "models", "core",
-                                      "llama")
-    dtype = 'float16'
-    cmodel_dir = os.path.join(engine_dir, "llama-7b-cmodel")
-
-    ckpt_dir = convert_weights(
-        llm_venv=llm_venv,
-        example_root=llama_example_root,
-        cmodel_dir=cmodel_dir,
-        model='llama-7b',
-        model_path=f'{llm_models_root()}/llama-models/llama-7b-hf',
-        data_type=dtype)
-
-    venv_check_call(llm_venv, [
-        run_llm_path,
-        "--model_dir",
-        ckpt_dir,
-    ])
-
-
-def test_llmapi_exit(llm_venv):
-    llm_exit_script = unittest_path() / "llmapi/run_llm_exit.py"
-    llama_model_dir = Path(
-        llm_models_root()) / "llama-models-v2/TinyLlama-1.1B-Chat-v1.0"
-
-    run_command = [
-        str(llm_exit_script), "--model_dir",
-        str(llama_model_dir), "--tp_size", "1"
-    ]
-    venv_check_call(llm_venv, run_command)
-
-
-@pytest.mark.skip_less_device(2)
-def test_llmapi_exit_multi_gpu(llm_venv):
-    llm_exit_script = unittest_path() / "llmapi/run_llm_exit.py"
-    llama_model_dir = Path(
-        llm_models_root()) / "llama-models-v2/TinyLlama-1.1B-Chat-v1.0"
-
-    run_command = [
-        str(llm_exit_script), "--model_dir",
-        str(llama_model_dir), "--tp_size", "2"
-    ]
-    venv_check_call(llm_venv, run_command)
-
-
-def test_llmapi_chat_example(llm_root, llm_venv):
-    # Test for the examples/apps/chat.py
-    example_root = Path(os.path.join(llm_root, "examples", "apps"))
-    test_root = unittest_path() / "llmapi" / "apps"
-    llm_venv.run_cmd([
-        "-m", "pip", "install", "-r",
-        os.path.join(example_root, "requirements.txt")
-    ])
-
-    llm_venv.run_cmd(["-m", "pytest", str(test_root / "_test_llm_chat.py")])
-
-
-def test_llmapi_server_example(llm_root, llm_venv):
-    # Test for the examples/apps/fastapi_server.py
-    test_root = unittest_path() / "llmapi" / "apps"
-    llm_venv.run_cmd(["-m", "pytest", str(test_root / "_test_llm_server.py")])
-
-
 def test_trtllm_serve_example(llm_root, llm_venv):
     example_root = Path(os.path.join(llm_root, "examples", "serve"))
     test_root = unittest_path() / "llmapi" / "apps"
@@ -1399,145 +1189,6 @@ def test_build_time_benchmark_sanity(llm_root, llm_venv):
         '-m',
         temp.name,
     ])
-
-
-### LLMAPI examples
-def _run_llmapi_example(llm_root, engine_dir, llm_venv, script_name: str,
-                        *args):
-    example_root = Path(llm_root) / "examples" / "llm-api"
-    engine_dir = Path(engine_dir) / "llmapi"
-    if not engine_dir.exists():
-        engine_dir.mkdir(parents=True)
-    examples_script = example_root / script_name
-
-    run_command = [str(examples_script)] + list(args)
-
-    # Create llm models softlink to avoid duplicated downloading for llm api example
-    src_dst_dict = {
-        f"{llm_models_root()}/llama-models-v2/TinyLlama-1.1B-Chat-v1.0":
-        f"{llm_venv.get_working_directory()}/TinyLlama/TinyLlama-1.1B-Chat-v1.0",
-        f"{llm_models_root()}/vicuna-7b-v1.3":
-        f"{llm_venv.get_working_directory()}/lmsys/vicuna-7b-v1.3",
-        f"{llm_models_root()}/medusa-vicuna-7b-v1.3":
-        f"{llm_venv.get_working_directory()}/FasterDecoding/medusa-vicuna-7b-v1.3",
-        f"{llm_models_root()}/llama3.1-medusa-8b-hf_v0.1":
-        f"{llm_venv.get_working_directory()}/nvidia/Llama-3.1-8B-Medusa-FP8",
-    }
-
-    for src, dst in src_dst_dict.items():
-        if not os.path.islink(dst):
-            os.makedirs(os.path.dirname(dst), exist_ok=True)
-            os.symlink(src, dst, target_is_directory=True)
-
-    cnn_dailymail_src = f"{llm_models_root()}/datasets/cnn_dailymail"
-    cnn_dailymail_dst = f"{llm_venv.get_working_directory()}/cnn_dailymail"
-    if not os.path.islink(cnn_dailymail_dst):
-        os.symlink(cnn_dailymail_src,
-                   cnn_dailymail_dst,
-                   target_is_directory=True)
-
-    venv_check_call(llm_venv, run_command)
-
-
-def test_llmapi_quickstart(llm_root, engine_dir, llm_venv):
-    _run_llmapi_example(llm_root, engine_dir, llm_venv, "quickstart_example.py")
-
-
-def test_llmapi_example_inference(llm_root, engine_dir, llm_venv):
-    _run_llmapi_example(llm_root, engine_dir, llm_venv, "llm_inference.py")
-
-
-def test_llmapi_example_customize(llm_root, engine_dir, llm_venv):
-    _run_llmapi_example(llm_root, engine_dir, llm_venv,
-                        "llm_inference_customize.py")
-
-
-def test_llmapi_example_inference_async(llm_root, engine_dir, llm_venv):
-    _run_llmapi_example(llm_root, engine_dir, llm_venv,
-                        "llm_inference_async.py")
-
-
-def test_llmapi_example_inference_async_streaming(llm_root, engine_dir,
-                                                  llm_venv):
-    _run_llmapi_example(llm_root, engine_dir, llm_venv,
-                        "llm_inference_async_streaming.py")
-
-
-def test_llmapi_example_quantization(llm_root, engine_dir, llm_venv):
-    _run_llmapi_example(llm_root, engine_dir, llm_venv, "llm_quantization.py")
-
-
-def test_llmapi_example_logits_processor(llm_root, engine_dir, llm_venv):
-    _run_llmapi_example(llm_root, engine_dir, llm_venv,
-                        "llm_logits_processor.py")
-
-
-def test_llmapi_example_multilora(llm_root, engine_dir, llm_venv):
-    _run_llmapi_example(llm_root, engine_dir, llm_venv, "llm_multilora.py")
-
-
-def test_llmapi_example_guided_decoding(llm_root, engine_dir, llm_venv):
-    _run_llmapi_example(llm_root, engine_dir, llm_venv,
-                        "llm_guided_decoding.py")
-
-
-def test_llmapi_example_lookahead_decoding(llm_root, engine_dir, llm_venv):
-    _run_llmapi_example(llm_root, engine_dir, llm_venv,
-                        "llm_lookahead_decoding.py")
-
-
-def test_llmapi_example_medusa_decoding(llm_root, engine_dir, llm_venv):
-    _run_llmapi_example(llm_root, engine_dir, llm_venv,
-                        "llm_medusa_decoding.py")
-
-
-def test_llmapi_example_medusa_decoding_use_modelopt(llm_root, engine_dir,
-                                                     llm_venv):
-    _run_llmapi_example(
-        llm_root, engine_dir, llm_venv, "llm_medusa_decoding.py",
-        "--use_modelopt_ckpt",
-        f"--model_dir={llm_models_root()}/llama3.1-medusa-8b-hf_v0.1")
-
-
-def test_llmapi_example_eagle_decoding(llm_root, engine_dir, llm_venv):
-    _run_llmapi_example(llm_root, engine_dir, llm_venv, "llm_eagle_decoding.py")
-
-
-@pytest.mark.skip_less_device(2)
-def test_llmapi_example_distributed_tp2(llm_root, engine_dir, llm_venv):
-    _run_llmapi_example(llm_root, engine_dir, llm_venv,
-                        "llm_inference_distributed.py")
-
-
-@pytest.mark.skip_less_device(2)
-def test_llmapi_example_distributed_autopp_tp2(llm_root, engine_dir, llm_venv):
-    _run_llmapi_example(llm_root, engine_dir, llm_venv, "llm_auto_parallel.py")
-
-
-def test_llmapi_quickstart_atexit(llm_root, engine_dir, llm_venv):
-    script_path = Path(
-        llm_root
-    ) / "tests/integration/defs/examples/run_llm_quickstart_atexit.py"
-    llm_venv.run_cmd([str(script_path)])
-
-
-def test_llmapi_quant_llama_70b(llm_root, engine_dir, llm_venv):
-    # Test quantizing llama-70b model with only 2 H100 GPUs
-    # The background: there is a bug preventing quantization of llama-70b model with <tp-size> GPUs
-    visible_devices = os.environ.get("CUDA_VISIBLE_DEVICES", "").split(',')
-    if len(visible_devices) < 2:
-        visible_devices = ['0', '1']
-    visible_devices = visible_devices[:2]
-
-    env = {
-        'CUDA_VISIBLE_DEVICES': ','.join(visible_devices),
-    }
-    print(f'env: {env}')
-
-    script_path = Path(
-        llm_root
-    ) / "tests/integration/defs/examples/run_llm_fp8_quant_llama_70b.py"
-    llm_venv.run_cmd([str(script_path)], env=env)
 
 
 # End of HLAPI examples
