+++ tests/integration/test_lists/test-db/l0_a10.yml
@@ -62,13 +62,13 @@ l0_a10:
   - unittest/trt/model/test_mistral.py
   - unittest/trt/model/test_llama.py
   - test_e2e.py::test_gpt3_175b_1layers_build_only # 6 mins
-  - test_e2e.py::test_llmapi_load_engine_from_build_command[llama-llama-models/llama-7b-hf] # 5min
-  - test_e2e.py::test_llmapi_build_command_parameters_align[llama-llama-models-v2/TinyLlama-1.1B-Chat-v1.0]
-  - test_e2e.py::test_llmapi_load_engine_from_build_command_with_lora[llama-llama-models-v2/llama-v2-7b-hf]
-  - test_e2e.py::test_llmapi_example_quantization
-  - test_e2e.py::test_llmapi_chat_example
-  - test_e2e.py::test_llmapi_exit
-  - test_e2e.py::test_llmapi_server_example
+  - llmapi/test_llm_e2e.py::test_llmapi_load_engine_from_build_command[llama-llama-models/llama-7b-hf] # 5min
+  - llmapi/test_llm_e2e.py::test_llmapi_build_command_parameters_align[llama-llama-models-v2/TinyLlama-1.1B-Chat-v1.0]
+  - llmapi/test_llm_e2e.py::test_llmapi_load_engine_from_build_command_with_lora[llama-llama-models-v2/llama-v2-7b-hf]
+  - llmapi/test_llm_examples.py::test_llmapi_example_quantization
+  - llmapi/test_llm_examples.py::test_llmapi_chat_example
+  - llmapi/test_llm_e2e.py::test_llmapi_exit
+  - llmapi/test_llm_examples.py::test_llmapi_server_example
   - test_e2e.py::test_trtllm_serve_example
   - test_e2e.py::test_trtllm_serve_multimodal_example
   - test_e2e.py::test_openai_misc_example
@@ -82,7 +82,7 @@ l0_a10:
   - accuracy/test_cli_flow.py::TestLlama7B::test_streamingllm # 2 mins
   - unittest/trt/functional # 37 mins
   - test_cache.py::test_cache_sanity # 1 sec
-  - test_e2e.py::test_llmapi_quickstart_atexit
+  - llmapi/test_llm_examples.py::test_llmapi_quickstart_atexit
   - unittest/api_stability
   - unittest/bindings
   - unittest/test_model_runner_cpp.py
@@ -157,8 +157,8 @@ l0_a10:
   - accuracy/test_cli_flow.py::TestPhi2::test_auto_dtype # 2 mins
   - accuracy/test_cli_flow.py::TestGemma2_9BIt::test_auto_dtype
   - examples/test_mamba.py::test_llm_mamba_1gpu[mamba-130m-float16-enable_gemm_plugin] # 2 mins
-  - test_e2e.py::test_llmapi_load_engine_from_build_command[llama-codellama/CodeLlama-7b-Instruct-hf] # 5min
-  - test_e2e.py::test_llmapi_load_ckpt_from_convert_command # 5min
+  - llmapi/test_llm_e2e.py::test_llmapi_load_engine_from_build_command[llama-codellama/CodeLlama-7b-Instruct-hf] # 5min
+  - llmapi/test_llm_e2e.py::test_llmapi_load_ckpt_from_convert_command # 5min
   - test_e2e.py::test_benchmark_sanity[t5_base]
   - examples/test_openai.py::test_llm_openai_triton_1gpu
   - examples/test_openai.py::test_llm_openai_triton_plugingen_1gpu
