+++ tests/integration/test_lists/test-db/l0_gb203.yml
@@ -28,11 +28,11 @@ l0_gb203:
   - examples/test_llama.py::test_llm_llama_1gpu[llama-3.1-8b-instruct-hf-fp8-enable_fp8-float16-summarization-nb:1]
   # - examples/test_qwen.py::test_llm_qwen1_5_7b_single_gpu_lora[qwen1.5_7b_chat-Qwen1.5-7B-Chat-750Mb-lora] # https://nvbugs/5234573
   # - examples/test_qwen.py::test_llm_qwen_single_gpu_summary[qwen2.5_1.5b_instruct-enable_paged_kv_cache-enable_remove_input_padding-enable_weight_only-enable_fmha_fp32_acc] # https://nvbugs/5234573
-  - test_e2e.py::test_llmapi_quickstart
-  - test_e2e.py::test_llmapi_example_inference
-  - test_e2e.py::test_llmapi_example_inference_async
-  - test_e2e.py::test_llmapi_example_inference_async_streaming
-  - test_e2e.py::test_llmapi_example_logits_processor
-  - test_e2e.py::test_llmapi_example_multilora
-  - test_e2e.py::test_llmapi_example_guided_decoding
-  - test_e2e.py::test_llmapi_example_customize
+  - llmapi/test_llm_examples.py::test_llmapi_quickstart
+  - llmapi/test_llm_examples.py::test_llmapi_example_inference
+  - llmapi/test_llm_examples.py::test_llmapi_example_inference_async
+  - llmapi/test_llm_examples.py::test_llmapi_example_inference_async_streaming
+  - llmapi/test_llm_examples.py::test_llmapi_example_logits_processor
+  - llmapi/test_llm_examples.py::test_llmapi_example_multilora
+  - llmapi/test_llm_examples.py::test_llmapi_example_guided_decoding
+  - llmapi/test_llm_examples.py::test_llmapi_example_customize
