+++ tests/integration/test_lists/test-db/l0_h100.yml
@@ -132,8 +132,8 @@ l0_h100:
   - unittest/test_model_runner_cpp.py
   - test_cache.py::test_cache_sanity # 1 sec
   - unittest/llmapi/test_llm_quant.py # 5.5 mins on H100
-  - test_e2e.py::test_llmapi_quickstart_atexit
   - test_e2e.py::test_mistral_large_hidden_vocab_size
+  - llmapi/test_llm_examples.py::test_llmapi_quickstart_atexit
   - examples/test_medusa.py::test_llm_medusa_with_qaunt_base_model_1gpu[fp8-use_py_session-medusa-vicuna-7b-v1.3-4-heads-float16-bs1]
   - examples/test_medusa.py::test_llm_medusa_with_qaunt_base_model_1gpu[fp8-use_cpp_session-medusa-vicuna-7b-v1.3-4-heads-float16-bs1]
   - unittest/trt/attention/test_gpt_attention_IFB.py
